requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python-headless
numpy
Pillow
timm